Rename buff to effect and fill Archer data

+++ IdleHeroes.Data/DTO/Actions/DamageActionDto.cs
@@ -2,7 +2,7 @@
 {
     public class DamageActionDto : ActionDto
     {
-        public double Potency { get; set; }
+        public int Potency { get; set; }
 
         public override void CreateAction(IActionDataBuilder builder)
         {
